Edit DetailPage -학습탭 중복 클릭시 컨텐츠 안보임 현상 수정
