+++ src/global.css
@@ -2,6 +2,7 @@
     width: 100%;
     height: 95%;
     box-sizing: border-box;
+    overflow-y: auto;
     display: flex;
     flex-direction: column;
 }
